feat: create new file and refactor
