+++ src/encoded/static/components/static-pages/placeholders/index.js
@@ -30,6 +30,7 @@ import { DownloadAllFilesFromSearchHrefButton } from '../components/challenges/D
 import { ChallengeTableWrapper } from '../components/challenges/ChallengeTables';
 import DataMatrix from '../../viz/Matrix/DataMatrix';
 import RetractedFilesTable from '../components/RetractedFilesTable';
+import RenamedFilesTable from '../components/RenamedFilesTable';
 import { SubmissionDataDictionary } from '../SubmissionDataDictionary';
 import { DonorManifestDataDictionary } from '../DonorManifestDataDictionary';
 import { PipelineDocsNavigation } from '../PipelineDocsNavigation';
@@ -60,6 +61,7 @@ const placeholders = {
     DownloadAllFilesFromSearchHrefButton,
     ChallengeTableWrapper,
     RetractedFilesTable,
+    RenamedFilesTable,
     DataMatrixComparisonTabs,
     DataMatrix,
     SubmissionDataDictionary,
